requirements.txt
Sanic
psycopg2-binary
sanic_openapi
sanic-cors